refactor: inline single-use pairing macros
